requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
sentence-transformers>=2.5.1
faiss-cpu>=1.7.4
google-generativeai>=0.7.0
numpy>=1.26.0
pydantic>=2.5.0
streamlit>=1.34.0

